Start redshift 15 seconds after i3 startup
This is an ugly hack that will manually fetch the approximate location
based on the IP address and pass these coordinates to redshift. It needs
to be delayed for several seconds to give enough time for a network
connection to be established. This is done to work around redshift's
apparent inability to connect to geoclue to get the location
automatically.